requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
<<<<<<< HEAD
MarkupSafe==3.0.2
=======
MarkupSafe==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.28
Flask-Migrate==4.0.5
python-dotenv==1.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
requests==2.31.0
>>>>>>> main
